WIP: Project

+++ project/src/kernal.c
@@ -4,41 +4,64 @@
 #include <string.h>
 #include <unistd.h>
 #include <stdbool.h>
+#include <sys/msg.h>
 #include "msgq.h"
 
 #define LIMIT 10
 
 void handler(int sig);
-unsigned int availableSize();
-int set(char* data);
-bool del(int pos);
 int up;
 int down;
-int CLK = 0;
-int size;
-unsigned int add_time = -1;
-unsigned int del_time = -1;
-char storage[LIMIT][BUFFER_SIZE];
-bool busy[LIMIT] = { false };
-msgt message_t;
-msgs message_s;
 
-void main()
+int main()
 {
+  message msg;
   up = getUpQueue();
   down = getDownQueue();
-  message_s = msgs_default;
 
-  signal(SIGUSR1, handler);
-  signal(SIGUSR2, handler);
+  signal(SIGALRM, handler);
+  alarm(1);
 
   while(true) {
-    killpg(getpgrp(), SIGUSR2);
-    sleep(4);
+    int size = msgrcv(up, &msg, sizeof(msg.content), KERNAL_ADDRESS, 0);
+
+    switch(msg.content.mtype) {
+      case DISK_ADDRESS:
+        switch (msg.content.mtext[0]) {
+          case ADD_DISK_SUCCESS:
+            break;
+          case ADD_DISK_FAILURE:
+            break;
+          case DEL_DISK_SUCCESS:
+            break;
+          case DEL_DISK_FAILURE:
+            break;
+          case SIZE_DISK_RESPONSE:
+            msg.content.mtext[1];
+            break;
+          default: break;
+        }
+        break;
+      default:
+        switch (msg.content.mtext[0]) {
+          case ADD_REQUEST:
+            break;
+          case DEL_REQUEST:
+            break;
+          default: break;
+        }
+        break;
+    }
   }
+  return 0;
 }
 
 void handler(int sig)
 {
-  printf("DEBUG: KERNAL\n");
+  if (sig == SIGALRM) {
+    killpg(getpgrp(), SIGUSR2);
+    alarm(1);
+  }
+
+  signal(SIGALRM, handler);
 }
